Add schedule block (unfinished)
To finish this schedule block will require a redesign of the data format, because as it is right now, I won't be able to select individual timings and edit them if I want to recombine them later, which I must (as required by the format). For that reason, this will currently remain on pause.